+++ example/lib/application/notes/notes_loading_blocs.dart
@@ -3,12 +3,12 @@ import 'package:example/application/core/blocs/action_blocs.dart';
 import 'package:example/domain/notes/facades/notes_facade.dart';
 import 'package:example/domain/notes/models/note.dart';
 
-class NotesCubit extends ReactiveFacadeConsumerCubit<PaginationData<Note>> {
+class NotesCubit extends ReactiveFacadeConsumerCubit<PaginationInfo<Note>> {
   NotesCubit(super.facade);
 }
 
 typedef NotesCubitState
-    = ReactiveFacadeConsumerCubitState<PaginationData<Note>>;
+    = ReactiveFacadeConsumerCubitState<PaginationInfo<Note>>;
 
 class NotesLoaderBloc extends VoidActionBloc<NotesLoadRequestedEvent> {
   NotesLoaderBloc(NotesFacade notesFacade)